python-requirements.txt
# Python Requirements for AI Integration Course Backend and Tools
# Install with: pip install -r python-requirements.txt

# Web Framework (Backend)
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Data & Validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI & ML
openai>=1.10.0
# sentence-transformers removed to avoid heavy torch/CUDA deps for this deploy
numpy>=1.26.0

# Google Cloud
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0

# Database
psycopg[binary]>=3.1.0

# Utilities
httpx>=0.26.0
youtube-transcript-api>=0.6.0
